requirements.txt
python-dotenv>=1.0.0
spacy>=3.0.0
requests>=2.31.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
alembic>=1.12.1
pytest>=7.4.3
O365==2.0.26
msal>=1.25.0
azure-ai-formrecognizer>=3.2.1
azure-identity>=1.15.0
azure-core>=1.29.5
azure-storage-blob>=12.19.0